Introduce a new class STANDARD and use it to install some new packages.
The intent is that normal image generation will enable this class. Packages
installed by this class are typically installed on a modern Linux cloud system
(this Priority: important and Priority: standard), but are not necessary for
basic functionality. Specialized derivative images may elect to leave this
class disabled in order to generate a minimal image with a smaller footprint.