Re-define version to include pipeline iid always
The version is now defined to contain two parts, the build id and the
pipeline iid.  The build id is either the date (for official builds) or
the namespace.  The pipeline iid will change for every build, so is used
to distinguish between different builds within the same build id.

Those two numbers also fit into the Azure version spec.