Install haveged on arm64 buster images
Arm64 does not support an RDRAND-like mechanism for obtaining entropy from
underlying hardware.  This change adds the haveged package to the image, which
provides a source of entropy based on timing measurements of various hardware
resources.

Linux 5.4 provides similar functionality within the kernel, so haveged is only
needed on buster at this time.