Support architecture dependent availability of the cloud kernel
The kernel is adding support for a cloud flavour for arm64 in addition to the
currently supported amd64.  This will be available in sid and bullseye, but
never buster, stretch, etc.  But the amd64 cloud kernel does exist in buster,
so we need to take both release and architecture into account when deciding
between the LINUX_IMAGE_CLOUD and LINUX_IMAGE_BASE classes.