Call dist-upgrade in hooks/updatebase
We skip FAI's built-in updatebase task in order to handle mount points
ourselves.  This means that we need to also take care to perform the
rest of the tasks that are expected from updatebase.  The most notable
is dist-upgrade.  Without this, we never apply security updates to
packages installed during the deboostrap phase of the installation.