aws: ssm: Ensure that we only set a given key once
In cases where the release name and release ID are identical in the
manifest (e.g. for 'sid'), we could end trying to set a variable more
than once.  In cases where overwrite=False, this would lead to an
exception and abort execution.  Handle this situation by tracking the
names of SSM variables as we set them, and avoid setting ones that
have already been seen.