Install libnss-myhostname
As observed in #1064235, cloud instances with neither libnss-resolve nor
libnss-myhostname are unable to resolve their own hostname.  On bookworm, this
causes issues as we recently started removing libnss-resolve due to name
resolution issues.  But even on newer releases, the libnss-resolve
documentation recommends installing both libnss-resolve and libnss-myhostname
side-by-side, e.g. from nss-resolve(8):

This duplicates the functionality of nss-myhostname(8), but it is still
recommended (see examples below) to keep nss-myhostname configured in
/etc/nsswitch.conf, to keep those names resolveable if systemd-resolved is not
running.